Add rain effect with render texture and camera
Introduced a new rain GameObject and a dedicated rAIN CAMERA to the Main Menu scene, including associated scripts and audio for thunder and lightning effects. Added a new render texture (rAIN.renderTexture) and updated the RainDrops material to use this texture with adjusted tiling and offset. Updated scene and camera settings to integrate the rain effect.